requirements/requirements_dev.txt
build
coreapi
django-debug-toolbar==3.2.4
django-test-migrations
drf-yasg
# pprofile - re-add once https://github.com/vpelletier/pprofile/issues/41 is addressed
ipython>=7.31.1 # https://github.com/ansible/awx/security/dependabot/30
unittest2
black
pytest!=7.0.0
pytest-asyncio
pytest-cov
pytest-django
pytest-mock
pytest-timeout
pytest-xdist
tox  # for awxkit
logutils
jupyter
# matplotlib - Caused issues when bumping to setuptools 58
backports.tempfile  # support in unit tests for py32+ tempfile.TemporaryDirectory
git+https://github.com/artefactual-labs/mockldap.git@master#egg=mockldap
gprof2dot
atomicwrites
flake8
yamllint
pip>=21.3,<=24.0 # PEP 660 – Editable installs for pyproject.toml based builds (wheel based)
python-ldap
# python debuggers
debugpy
remote-pdb
sdb
PyJWT